+++ lib/shiika/program/env.rb
@@ -95,24 +95,41 @@ module Shiika
         return ivar
       end
 
-      def find_method(receiver_type, name)
-        case receiver_type
-        when TyRaw, TyMeta
-          sk_class = @data[:sk_classes].fetch(receiver_type.name)
-          return sk_class.find_method(name)
-        when TySpe, TySpeMeta
-          gen_cls = @data[:sk_classes].fetch(receiver_type.base_name)
-          sp_cls = gen_cls.specialized_class(receiver_type.type_args, self)
-          return sp_cls.find_method(name)
-        when TyGenMeta  # Class method of a generic class
-          gen_meta_cls = @data[:sk_classes].fetch(receiver_type.base_meta_name)
-          return gen_meta_cls.find_method(name)
-        when TyParam
-          typaram = receiver_type
-          return find_method(typaram.upper_bound, name)
+      def find_method(receiver_type, name, origin: receiver_type)
+        sk_method = 
+          case receiver_type
+          when TyRaw, TyMeta
+            sk_class = @data[:sk_classes].fetch(receiver_type.name)
+            sk_class.find_method(name)
+          when TySpe, TySpeMeta
+            gen_cls = @data[:sk_classes].fetch(receiver_type.base_name)
+            sp_cls = gen_cls.specialized_class(receiver_type.type_args, self)
+            sp_cls.find_method(name)
+          when TyGenMeta  # Class method of a generic class
+            gen_meta_cls = @data[:sk_classes].fetch(receiver_type.base_meta_name)
+            gen_meta_cls.find_method(name)
+          when TyParam
+            typaram = receiver_type
+            find_method(typaram.upper_bound, name)
+          else
+            raise "env.find_method(#{receiver_type}) is not implemented yet"
+          end
+        if sk_method
+          return sk_method
         else
-          raise "env.find_method(#{receiver_type}) is not implemented yet"
+          if (super_type = supertype_of(receiver_type))
+            return find_method(super_type, name, origin: origin)
+          else
+            raise SkTypeError, "method `#{name}' not found on #{origin}"
+          end
         end
+      end
+
+      def supertype_of(type)
+        sk_class = find_class(type.name)
+        return nil if sk_class.superclass_name == '__noparent__'
+        super_class = find_class(sk_class.superclass_name)
+        return super_class.type
       end
 
       def sk_self